broker/requirements.txt
pyzmq
influxdb
blosc
pandas
numpy
scipy
spectrum
tornado